+++ src/main/resources/application.properties
@@ -7,4 +7,3 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/desi_2023?serverTimezone=UTC
 spring.datasource.username=root
 spring.datasource.password=tsti_DESI_2023
-
